fix: Navigation theme colors
- `NavigationThemeProvider` doesn't seem to work anymore.
- Also only apply `headerShown: false` to the `HomeScreens` instead of everything and then reverting via manually assigning `headerShown: true`.
